Move ParseBackendSearch into search.go
