+++ pkg/api/search.go
@@ -1,0 +1,61 @@
+package api
+
+import (
+	"errors"
+	"fmt"
+	"net/http"
+	"strconv"
+)
+
+const (
+	URLParamLimit = "limit"
+	urlParamStart = "start"
+	urlParamEnd   = "end"
+
+	// todo(search): make configurable
+	maxRange     = 1800 // 30 minutes
+	defaultLimit = 20
+)
+
+func ParseBackendSearch(r *http.Request) (start, end int64, limit int, err error) {
+	if s := r.URL.Query().Get(urlParamStart); s != "" {
+		start, err = strconv.ParseInt(s, 10, 64)
+		if err != nil {
+			return
+		}
+	}
+
+	if s := r.URL.Query().Get(urlParamEnd); s != "" {
+		end, err = strconv.ParseInt(s, 10, 64)
+		if err != nil {
+			return
+		}
+	}
+
+	if s := r.URL.Query().Get(URLParamLimit); s != "" {
+		limit, err = strconv.Atoi(s)
+		if err != nil {
+			return
+		}
+	}
+
+	if start == 0 || end == 0 {
+		err = errors.New("please provide non-zero values for http parameters start and end")
+		return
+	}
+
+	if limit == 0 {
+		limit = defaultLimit
+	}
+
+	if end-start > maxRange {
+		err = fmt.Errorf("range specified by start and end exceeds %d seconds. received start=%d end=%d", maxRange, start, end)
+		return
+	}
+	if end <= start {
+		err = fmt.Errorf("http parameter start must be before end. received start=%d end=%d", start, end)
+		return
+	}
+
+	return
+}
